qdev, spapr: Add alias for CPU socket device
device_add semantics for CPU hotplug is still being discussed and developed
in the QEMU community. Since the interface hasn't been frozen yet, we
want to expose an interface that is very sPAPR specific so that in future
we could switch over to whatever generic semantics that evolves. In order
to support this, introduce an alias "spapr-cpu-socket" that should be used
to hot add and remove CPUs in sPAPR.

Currently for sPAPR, we support full hotplug of a socket at once. Socket
is a container of CPU cores and in PowerPC, it is usually one core per
socket while it is possible to define and boot a guest with multiple cores
per socket. In the current implementation number of cores per socket is
always fixed at boot time and any hotplugged socket will come with those
many number of cores only. i,e., there is no provision to add cores
incrementally to an existing socket. All the cores of a socket are added
at once or none are added.

Note on Migration: The current CPU hotplug implementation supports the
migration of VM which has undergone CPU hotplug and it supports the
hot removal of that CPU from the target after migration. However the
CPUs have to be removed in the order they have been added (LIFO) since
the current semantics doesn't provide a way to easily handle removal
of CPUs arbitrarily and still support VM migration.

NOTE: Not for upstream.

Signed-off-by: Bharata B Rao <bharata@linux.vnet.ibm.com>